fix(auth): split pages into server components for generateMetadata
