+++ app/[locale]/(auth)/reset-password/page.tsx
@@ -1,11 +1,25 @@
-"use client"
-
 import { Suspense } from "react"
-import { useTranslations } from "next-intl"
+import { getTranslations } from "next-intl/server"
+import { Metadata } from "next"
+
 import { ResetPasswordForm } from "./reset-password-form"
 
-function LoadingState() {
-  const t = useTranslations()
+export async function generateMetadata({
+  params,
+}: {
+  params: Promise<{ locale: string }>
+}): Promise<Metadata> {
+  const { locale } = await params
+  const t = await getTranslations({ locale, namespace: "auth.resetPassword" })
+
+  return {
+    title: t("title"),
+    description: t("description"),
+  }
+}
+
+async function LoadingState() {
+  const t = await getTranslations()
 
   return (
     <div className="flex min-h-screen items-center justify-center bg-background p-4">
