+++ app/[locale]/(auth)/sign-up/page.tsx
@@ -1,25 +1,7 @@
-"use client"
-
-import { useState } from "react"
 import { getTranslations } from "next-intl/server"
 import { Metadata } from "next"
-import { useTranslations } from "next-intl"
-import { Github, Loader2 } from "lucide-react"
 
-import { Link, useRouter } from "@/lib/navigation"
-import { Button } from "@/components/ui/button"
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card"
-import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
-import { authClient } from "@/lib/auth-client"
-import { useSignUpSchema } from "@/lib/schemas"
+import { SignUpForm } from "./sign-up-form"
 
 export async function generateMetadata({
   params,
@@ -36,205 +18,5 @@ export async function generateMetadata({
 }
 
 export default function SignUpPage() {
-  const t = useTranslations()
-  const router = useRouter()
-  const signUpSchema = useSignUpSchema()
-  const [name, setName] = useState("")
-  const [email, setEmail] = useState("")
-  const [password, setPassword] = useState("")
-  const [confirmPassword, setConfirmPassword] = useState("")
-  const [isLoading, setIsLoading] = useState(false)
-  const [errors, setErrors] = useState<Record<string, string>>({})
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault()
-    setIsLoading(true)
-    setErrors({})
-
-    const result = signUpSchema.safeParse({
-      name,
-      email,
-      password,
-      confirmPassword,
-    })
-
-    if (!result.success) {
-      const fieldErrors: Record<string, string> = {}
-      result.error.issues.forEach(
-        (err: { path: (string | number | symbol)[]; message: string }) => {
-          if (err.path[0]) {
-            fieldErrors[err.path[0] as string] = err.message
-          }
-        }
-      )
-      setErrors(fieldErrors)
-      setIsLoading(false)
-      return
-    }
-
-    await authClient.signUp.email(
-      { email, password, name },
-      {
-        onSuccess: () => {
-          setIsLoading(false)
-          router.push("/dashboard")
-        },
-        onError: (ctx) => {
-          setErrors({
-            form: t("auth.signUp.failedToSignUp"),
-          })
-          setIsLoading(false)
-        },
-      }
-    )
-  }
-
-  const handleOAuthSignIn = async (provider: "github" | "google") => {
-    await authClient.signIn.social({
-      provider,
-    })
-  }
-
-  return (
-    <div className="flex min-h-screen items-center justify-center bg-background p-4">
-      <Card className="w-full max-w-md">
-        <CardHeader className="space-y-1">
-          <CardTitle className="text-2xl font-bold">
-            {t("auth.signUp.title")}
-          </CardTitle>
-          <CardDescription>{t("auth.signUp.description")}</CardDescription>
-        </CardHeader>
-        <form onSubmit={handleSubmit}>
-          <CardContent className="space-y-4">
-            {errors.form && (
-              <div className="rounded-md bg-destructive/10 p-3 text-sm text-destructive">
-                {errors.form}
-              </div>
-            )}
-            <div className="space-y-2">
-              <Label htmlFor="name">{t("auth.signUp.nameLabel")}</Label>
-              <Input
-                id="name"
-                type="text"
-                placeholder={t("auth.signUp.namePlaceholder")}
-                value={name}
-                onChange={(e) => setName(e.target.value)}
-                disabled={isLoading}
-              />
-              {errors.name && (
-                <p className="text-sm text-destructive">{errors.name}</p>
-              )}
-            </div>
-            <div className="space-y-2">
-              <Label htmlFor="email">{t("common.email")}</Label>
-              <Input
-                id="email"
-                type="email"
-                placeholder={t("auth.signUp.emailPlaceholder")}
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-                disabled={isLoading}
-              />
-              {errors.email && (
-                <p className="text-sm text-destructive">{errors.email}</p>
-              )}
-            </div>
-            <div className="space-y-2">
-              <Label htmlFor="password">{t("auth.signUp.passwordLabel")}</Label>
-              <Input
-                id="password"
-                type="password"
-                placeholder={t("auth.signUp.passwordPlaceholder")}
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-                disabled={isLoading}
-              />
-              {errors.password && (
-                <p className="text-sm text-destructive">{errors.password}</p>
-              )}
-            </div>
-            <div className="space-y-2">
-              <Label htmlFor="confirmPassword">
-                {t("auth.signUp.confirmPasswordLabel")}
-              </Label>
-              <Input
-                id="confirmPassword"
-                type="password"
-                placeholder={t("auth.signUp.confirmPasswordPlaceholder")}
-                value={confirmPassword}
-                onChange={(e) => setConfirmPassword(e.target.value)}
-                disabled={isLoading}
-              />
-              {errors.confirmPassword && (
-                <p className="text-sm text-destructive">
-                  {errors.confirmPassword}
-                </p>
-              )}
-            </div>
-          </CardContent>
-          <CardFooter className="mt-6 flex flex-col space-y-4">
-            <Button type="submit" className="w-full" disabled={isLoading}>
-              {isLoading && <Loader2 className="mr-2 h-4 w-4 animate-spin" />}
-              {t("auth.signUp.signUpButton")}
-            </Button>
-            <div className="relative w-full">
-              <div className="absolute inset-0 flex items-center">
-                <span className="w-full border-t" />
-              </div>
-              <div className="relative flex justify-center text-xs uppercase">
-                <span className="bg-card px-2 text-muted-foreground">
-                  {t("common.orContinueWith")}
-                </span>
-              </div>
-            </div>
-            <div className="flex gap-2">
-              <Button
-                type="button"
-                variant="outline"
-                className="flex-1"
-                onClick={() => handleOAuthSignIn("github")}
-                disabled={isLoading}
-              >
-                <Github className="mr-2 h-4 w-4" />
-                GitHub
-              </Button>
-              <Button
-                type="button"
-                variant="outline"
-                className="flex-1"
-                onClick={() => handleOAuthSignIn("google")}
-                disabled={isLoading}
-              >
-                <svg className="mr-2 h-4 w-4" viewBox="0 0 24 24">
-                  <path
-                    d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"
-                    fill="#4285F4"
-                  />
-                  <path
-                    d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"
-                    fill="#34A853"
-                  />
-                  <path
-                    d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"
-                    fill="#FBBC05"
-                  />
-                  <path
-                    d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"
-                    fill="#EA4335"
-                  />
-                </svg>
-                Google
-              </Button>
-            </div>
-            <p className="text-center text-sm text-muted-foreground">
-              {t("auth.signUp.hasAccount")}{" "}
-              <Link href="/sign-in" className="text-primary hover:underline">
-                {t("auth.signUp.signInLink")}
-              </Link>
-            </p>
-          </CardFooter>
-        </form>
-      </Card>
-    </div>
-  )
+  return <SignUpForm />
 }
